Refactor suggestion base classes, add email method
