+++ controllers/suggestions/suggest_event_webcast_review_controller.py
@@ -6,7 +6,7 @@ from google.appengine.ext import ndb
 from google.appengine.ext.webapp import template
 
 from consts.account_permissions import AccountPermissions
-from controllers.suggestions.suggestions_review_base_controller import SuggestionsReviewBaseController
+from controllers.suggestions.suggestions_base_controller import SuggestionsReviewBaseController
 from helpers.event.event_webcast_adder import EventWebcastAdder
 from helpers.memcache.memcache_webcast_flusher import MemcacheWebcastFlusher
