+++ controllers/suggestions/suggest_review_home_controller.py
@@ -3,7 +3,7 @@ import os
 from google.appengine.ext.webapp import template
 
 from consts.account_permissions import AccountPermissions
-from controllers.suggestions.suggestions_review_base_controller import SuggestionsReviewBaseController
+from controllers.suggestions.suggestions_base_controller import SuggestionsReviewBaseController
 from helpers.suggestions.suggestion_fetcher import SuggestionFetcher
 from models.suggestion import Suggestion
