+++ controllers/suggestions/suggest_team_media_review_controller.py
@@ -7,7 +7,7 @@ from google.appengine.ext import ndb
 
 from consts.account_permissions import AccountPermissions
 from consts.media_type import MediaType
-from controllers.suggestions.suggestions_review_base_controller import SuggestionsReviewBaseController
+from controllers.suggestions.suggestions_base_controller import SuggestionsReviewBaseController
 from helpers.media_manipulator import MediaManipulator
 from helpers.suggestions.media_creator import MediaCreator
 from models.media import Media
