--- controllers/suggestions/suggestions_review_base_controller.py
@@ -1,17 +1,0 @@
-from controllers.base_controller import LoggedInHandler
-
-
-class SuggestionsReviewBaseController(LoggedInHandler):
-    """
-    Base controller for reviewing suggestions.
-    """
-
-    REQUIRED_PERMISSIONS = []
-
-    def __init__(self, *args, **kw):
-        super(SuggestionsReviewBaseController, self).__init__(*args, **kw)
-        self.verify_permissions()
-
-    def verify_permissions(self):
-        for permission in self.REQUIRED_PERMISSIONS:
-            self._require_permission(permission)
